add money transfer test add version control to cards add admin controller

+++ src/main/java/com/example/bankcards/entity/CardEntity.java
@@ -41,5 +41,5 @@ public class CardEntity {
 
     @Column(name = "version")
     @Version
-    private long version;
+    private int version;
 }
